// types/voteProxyContract.d.ts
export type VoteProxyContract = {
    getProxyAddress: () => string;
    getColdAddress: () => string;
    getHotAddress: () => string;
    lock: () => Promise<any>;
    free: () => Promise<any>;
    voteExec: (picks: string[] | string) => Promise<any>;
    getNumDeposits: () => Promise<any>;
    getVotedProposalAddresses: () => Promise<any>;
  };
  
  export type OldVoteProxyContract = {
    role: string;
    address: string;
  };